+++ v3/crates/metadata-resolve/src/stages/data_connectors/mod.rs
@@ -3,20 +3,22 @@ use crate::types::subgraph::Qualified;
 mod error;
 mod types;
 
-pub use error::DataConnectorError;
-pub use error::NamedDataConnectorError;
+pub use error::{
+    DataConnectorError, DataConnectorIssue, NamedDataConnectorError, NamedDataConnectorIssue,
+};
 use std::collections::BTreeMap;
 pub use types::{
     ArgumentPreset, CommandsResponseConfig, DataConnectorCapabilities, DataConnectorContext,
-    DataConnectorLink, DataConnectorSchema, DataConnectors, NdcVersion,
+    DataConnectorLink, DataConnectorSchema, DataConnectors, DataConnectorsOutput, NdcVersion,
 };
 
 /// Resolve data connectors.
 pub fn resolve<'a>(
     metadata_accessor: &'a open_dds::accessor::MetadataAccessor,
     configuration: &Configuration,
-) -> Result<types::DataConnectors<'a>, NamedDataConnectorError> {
+) -> Result<types::DataConnectorsOutput<'a>, NamedDataConnectorError> {
     let mut data_connectors = BTreeMap::new();
+    let mut issues = vec![];
     for open_dds::accessor::QualifiedObject {
         subgraph,
         object: data_connector,
@@ -25,15 +27,27 @@ pub fn resolve<'a>(
         let qualified_data_connector_name =
             Qualified::new(subgraph.to_string(), data_connector.name.clone());
 
-        let data_conntext_context =
+        let (data_connector_context, connector_issues) =
             types::DataConnectorContext::new(data_connector, &configuration.unstable_features)
                 .map_err(|error| NamedDataConnectorError {
                     data_connector_name: qualified_data_connector_name.clone(),
                     error,
                 })?;
 
+        issues.extend(
+            connector_issues
+                .into_iter()
+                .map(|issue| NamedDataConnectorIssue {
+                    data_connector_name: qualified_data_connector_name.clone(),
+                    issue,
+                }),
+        );
+
         if data_connectors
-            .insert(qualified_data_connector_name.clone(), data_conntext_context)
+            .insert(
+                qualified_data_connector_name.clone(),
+                data_connector_context,
+            )
             .is_some()
         {
             return Err(NamedDataConnectorError {
@@ -42,5 +56,8 @@ pub fn resolve<'a>(
             });
         }
     }
-    Ok(types::DataConnectors(data_connectors))
+    Ok(types::DataConnectorsOutput {
+        data_connectors: types::DataConnectors(data_connectors),
+        issues,
+    })
 }