+++ v3/crates/metadata-resolve/src/stages/mod.rs
@@ -42,7 +42,12 @@ pub fn resolve(
         graphql_config::resolve(&metadata_accessor.graphql_config, metadata_accessor.flags)?;
 
     // Fetch and check schema information for all our data connectors
-    let data_connectors = data_connectors::resolve(&metadata_accessor, &configuration)?;
+    let data_connectors::DataConnectorsOutput {
+        data_connectors,
+        issues,
+    } = data_connectors::resolve(&metadata_accessor, &configuration)?;
+
+    all_warnings.extend(issues.into_iter().map(Warning::from));
 
     // Validate object types defined in metadata
     let object_types::ObjectTypesOutput {